chain: extract Chain/ChainView traits; move Zaino impl to chain/zaino.rs
chain.rs is now the backend-neutral interface (zero zaino_* imports); the Zaino implementation lives in chain/zaino.rs and implements the traits. Stream methods return BoxStream<'_> rather than associated stream types, because the block/mempool streams borrow the pinned view and so cannot be 'static. A mock-based test locks the ChainView contract. Consumers still name the concrete ZainoChain pending the generic conversion.
